experimental moves_info refactoring

+++ include/calculate_base_damage.h
@@ -3,6 +3,6 @@
 
 #include "global.h"
 
-s32 CalculateBaseDamage(struct BattlePokemon *attacker, struct BattlePokemon *defender, u32 move, u16 sideStatus, u16 powerOverride, u8 typeOverride, u8 bankAtk, u8 bankDef);
+s32 CalculateBaseDamage(struct BattlePokemon *attacker, struct BattlePokemon *defender, u32 move, u32 sideStatus, u16 powerOverride, u8 typeOverride, u8 bankAtk, u8 bankDef);
 
 #endif // GUARD_CALCULATE_BASE_DAMAGE_H
